db support

+++ main.js
@@ -2,16 +2,31 @@ const http = require('http');
 const express = require('express');
 const child_process = require('child_process');
 const fs = require('fs');
+const common = require('./common');
+const wikialbums = require('./wikialbums');
 
 const app = express();
 const port = 3000;
 
-const pw = fs.readFileSync('/main/TOKEN_FILE', 'utf8');
-fs.unlinkSync('/main/TOKEN_FILE');
-process.env['SECRETS_ACCESS_TOKEN'] = pw;
+const ENABLE_VAULT = true;
 
-const configRaw = fs.readFileSync('/config.json', 'utf8');
-const config = JSON.parse(configRaw);
+let vaultConfig;
+if (ENABLE_VAULT) {
+    const pw = fs.readFileSync('/main/TOKEN_FILE', 'utf8');
+    fs.unlinkSync('/main/TOKEN_FILE');
+    process.env['SECRETS_ACCESS_TOKEN'] = pw;
+
+    const configRaw = fs.readFileSync('/config.json', 'utf8');
+    vaultConfig = JSON.parse(configRaw);
+}
+
+const dbProjects = {
+    'wikialbums': wikialbums
+};
+
+const Log = (l) => {
+    common.Log('main.js', l);
+};
 
 app.use(express.json());
 
@@ -19,13 +34,13 @@ app.get('/vaultSecret', (req, res) => {
     const queryParams = req.query;
     const queryKey = queryParams.key;
 
-    if (!(queryKey in config)) {
-        console.log("No key found: " + queryKey);
+    if (!(queryKey in vaultConfig)) {
+        Log("No key found: " + queryKey);
         return;
     }
 
-    const key = config[queryKey];
-    console.log('Fetching key: ' + key);
+    const key = vaultConfig[queryKey];
+    Log('Fetching key: ' + key);
 
     let secretKey = child_process.execSync(`secrets.sh ${key}`, { encoding: 'utf8' }).trim();
 
@@ -33,6 +48,29 @@ app.get('/vaultSecret', (req, res) => {
         message: 'Received GET request',
         data: queryParams,
         secret: secretKey
+    });
+});
+
+app.get('/dbAction', async (req, res) => {
+    const queryParams = req.query;
+    const queryProj = queryParams.proj;
+
+    if (!queryProj || !dbProjects[queryProj]) {
+        Log("No project found: " + queryProj);
+        return;
+    }
+
+    let response = await dbProjects[queryProj].handle(queryParams);
+    if (response) {
+        res.json(response);
+    }
+});
+
+app.get('/test', (req, res) => {
+    const bodyData = req.body;
+    res.json({
+        message: 'Received GET request',
+        data: bodyData
     });
 });
 
@@ -47,5 +85,6 @@ app.post('/', (req, res) => {
 
 // Start the server
 app.listen(port, () => {
-    console.log(`Server is listening on http://localhost:${port}`);
+    Log(`Server is listening on http://localhost:${port}`);
 });
+
